#1563 Reverted tg-ui-action changes
